# apps/disher-backend-3.0/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Disher API",
    "version": "1.0.0",
    "description": "Backend for the Disher food diary. Errors share one RFC 9457-subset problem+json shape — see api/errors.ts."
  },
  "components": {
    "securitySchemes": {
      "cookieSession": {
        "type": "apiKey",
        "in": "cookie",
        "name": "__Secure-disher1.session_token"
      }
    },
...
  },
  "paths": {
...
    "/api/free-text-food/parse": {
      "post": {
        "operationId": "parseFreeTextFood",
        "tags": [
          "free-text"
        ],
        "description": "Free-text meal phrase → catalog-matched items (resolved / ambiguous / unresolved). Paid.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "title": "FreeTextFoodParseRequest",
                "type": "object",
                "required": [
                  "text"
                ],
                "properties": {
                  "text": {
                    "type": "string"
                  },
                  "existingDishNames": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "required": [
                        "id",
                        "name"
                      ],
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "name": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "cookieSession": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
...
  }
}